Refactor component system

+++ JungleTechLab_Week3/Json/json.hpp
@@ -258,7 +258,7 @@ class JSON
             return Internal.List->at( index );
         }
 
-        int length() const {
+        size_t length() const {
             if( Type == Class::Array )
                 return Internal.List->size();
             else
@@ -271,7 +271,7 @@ class JSON
             return false;
         }
 
-        int size() const {
+		size_t size() const {
             if( Type == Class::Object )
                 return Internal.Map->size();
             else if( Type == Class::Array )
